Fix some warnings.

+++ usrsctplib/netinet/sctp_bsd_addr.c
@@ -830,7 +830,7 @@ sctp_get_mbuf_for_msg(unsigned int space_needed, int want_header,
 		mbuf_threshold = 1;
 
 
-	if (space_needed > (((mbuf_threshold - 1) * MLEN) + MHLEN)) {
+	if ((int)space_needed > (((mbuf_threshold - 1) * MLEN) + MHLEN)) {
 		MCLGET(m, how);
 		if (m == NULL) {
 			return (NULL);
